test(skills): split loader/tests.rs and catalog/tests.rs into per-feature submodules (#494)
Both files were over 1,000 LOC and copy-pasted their tempdir + SKILL.md setup across dozens of tests. Each is now a directory with one file per existing inline mod test_xxx { ... } block, plus a fixtures.rs that exposes a shared SkillTestFixture / make_test_catalog helpers.

loader/tests/ — 11 split files, largest 175 LOC (test_metadata_compat). 13 tests in test_parse_skill_md adopt SkillTestFixture::with_body / empty / write_file.

catalog/tests/ — 7 split files, largest 222 LOC (test_catalog_crud). The 7 files all share fixtures::make_test_catalog / make_test_skill / add_skill_with_scope.

Behavioural parity: 207 tests pass (same as the baseline before split). cargo fmt + cargo clippy -p dcc-mcp-skills --all-targets -- -D warnings clean.

The remaining offenders (registry/tests.rs 648 LOC, pipeline/tests.rs 598 LOC, validator/tests.rs 513 LOC) are deferred to a follow-up because they live in dcc-mcp-actions and need their own fixture design.

Closes #494 partially (skills crate); follow-up issue to be filed for dcc-mcp-actions test files.
