+++ crates/dcc-mcp-skills/src/catalog/tests/mod.rs
@@ -1,0 +1,10 @@
+use super::*;
+
+pub(super) mod fixtures;
+mod test_catalog_crud;
+mod test_dispatcher;
+mod test_execute_script;
+mod test_execute_script_env;
+mod test_resolve_tool_script;
+mod test_search_hint;
+mod test_search_skills;
